move contract logic to resolver
