+++ v2/perps-v2/ui/src/hooks/usePositions.ts
@@ -13,8 +13,9 @@ import {
   address as multiCallAddress,
   Multicall3,
 } from '@synthetixio/v3-contracts/build/optimism-mainnet/Multicall3';
-import { useQuery as useReactQuery } from '@tanstack/react-query';
-import Wei, { wei } from '@synthetixio/wei';
+import { wei } from '@synthetixio/wei';
+import { ContractData, SubgraphPositionData, PositionData } from '../types/positions';
+import { POSITIONS_CONTRACT_QUERY } from '../queries/resolved';
 
 export function notNill<Value>(value: Value | null | undefined): value is Value {
   return value !== null && value !== undefined;
@@ -24,105 +25,13 @@ const provider = new providers.InfuraProvider(10, infuraId);
 const contract = new Contract(address, abi, provider) as PerpsV2MarketData;
 const Multicall3Contract = new Contract(multiCallAddress, multiCallAbi, provider) as Multicall3;
 
-interface SubgraphPositionData {
-  market: string;
-  asset: string;
-  avgEntryPrice: Wei;
-  leverage: Wei;
-  fees: Wei;
-  pnlAtLastModification: Wei;
-  netFundingAtLastModification: Wei;
-  fillPriceAtLastInteraction: Wei;
-}
-
-interface ContractData {
-  skew: Wei;
-  skewScale: Wei;
-  indexPrice: Wei;
-  size: Wei;
-  liquidationPrice: Wei;
-  accessibleMargin: Wei;
-  accruedFundingSinceLastModification: Wei;
-}
-
-export const calculateMarkPrice = ({
-  skew,
-  indexPrice,
-  skewScale,
-}: {
-  skew: Wei;
-  indexPrice: Wei;
-  skewScale: Wei;
-}) => {
-  const skewRatio = skew.div(skewScale);
-  const markPrice = indexPrice.mul(wei(1).add(skewRatio));
-  return markPrice;
-};
-
-export const calculateNewPnl = (
-  subgraphPositionData: SubgraphPositionData,
-  contractData: ContractData,
-  marketPrice: Wei
-) => {
-  const priceShiftSinceModification = marketPrice.sub(
-    subgraphPositionData.fillPriceAtLastInteraction
-  );
-  const pnlSinceModification = contractData.size.mul(priceShiftSinceModification);
-  const newPnl = subgraphPositionData.pnlAtLastModification
-    .add(pnlSinceModification)
-    .add(contractData.accruedFundingSinceLastModification);
-  return newPnl;
-};
-
-export const calculatePnlPercentage = (
-  subgraphPositionData: SubgraphPositionData,
-  contractData: ContractData,
-  pnl: Wei
-) => {
-  if (pnl.eq(0)) return wei(0);
-
-  const initialValue = contractData.size.mul(subgraphPositionData.avgEntryPrice);
-  const currentVal = initialValue.add(pnl);
-  const shift = currentVal.sub(initialValue);
-  return shift.div(initialValue);
-};
-
-export const calculatePositionData = (
-  subgraphPositionData: SubgraphPositionData,
-  contractData: ContractData,
-  address?: string
-) => {
-  if (contractData.size.eq(0)) return null;
-  const marketPrice = calculateMarkPrice(contractData);
-  const pnl = calculateNewPnl(subgraphPositionData, contractData, marketPrice);
-  const notionalValue = contractData.size.mul(marketPrice);
-  const pnlPercentage = calculatePnlPercentage(subgraphPositionData, contractData, pnl);
-  const netFunding = subgraphPositionData.netFundingAtLastModification.add(
-    contractData.accruedFundingSinceLastModification
-  );
-
-  return {
-    asset: subgraphPositionData.asset,
-    indexPrice: contractData.indexPrice,
-    liquidationPrice: contractData.liquidationPrice,
-    pnl,
-    pnlPercentage,
-    margin: contractData.accessibleMargin,
-    size: contractData.size,
-    long: contractData.size.gt(0),
-    avgEntryPrice: subgraphPositionData.avgEntryPrice,
-    leverage: subgraphPositionData.leverage,
-    funding: netFunding,
-    marketPrice,
-    notionalValue: notionalValue,
-    fees: subgraphPositionData.fees,
-    address,
-  };
-};
-
 export const usePositions = (walletAddress?: string) => {
   // Initial query to give a list of markets
-  const { data, error } = useQuery(POSITIONS_QUERY_MARKET, {
+  const {
+    data: marketData,
+    loading: marketLoading,
+    error: marketError,
+  } = useQuery(POSITIONS_QUERY_MARKET, {
     variables: {
       where: { isOpen: true, account: walletAddress },
       orderBy: FuturesPosition_OrderBy.Size,
@@ -131,7 +40,8 @@ export const usePositions = (walletAddress?: string) => {
     },
     pollInterval: 5000,
   });
-  const openPositions = data?.futuresPositions.map((item) => ({
+
+  const openPositions = marketData?.futuresPositions.map((item) => ({
     market: item.market.marketKey,
     asset: item.market.asset,
     avgEntryPrice: wei(item.avgEntryPrice, 18, true),
@@ -142,35 +52,22 @@ export const usePositions = (walletAddress?: string) => {
     fillPriceAtLastInteraction: wei(item.lastPrice, 18, true),
   }));
 
-  return useReactQuery({
-    queryKey: [
-      'usePosition',
-      {
-        walletAddress,
-        subgraphData: JSON.stringify(data),
-      },
-    ],
-    queryFn: async () => {
-      if (!openPositions) throw Error('Query should not be enabled');
-      if (error) throw error;
-
-      const positionsData = await fetchPositions(openPositions, walletAddress || '');
-      return positionsData
-        .map((contractData, index) => {
-          const calculatedPositionData = calculatePositionData(
-            openPositions[index],
-            contractData,
-            walletAddress
-          );
-          return calculatedPositionData;
-        })
-        .filter(notNill);
-    },
-    enabled: Boolean(openPositions),
+  const { data, loading, error } = useQuery(POSITIONS_CONTRACT_QUERY, {
+    variables: { walletAddress, openPositions },
+    skip: marketData?.futuresPositions ? false : true,
+    pollInterval: 1000,
   });
+
+  const positionData = (data?.positionsFromContract as PositionData[]) || undefined;
+
+  return {
+    data: positionData,
+    loading: loading || marketLoading,
+    error: error || marketError,
+  };
 };
 
-async function fetchPositions(
+export async function fetchPositions(
   positionData: SubgraphPositionData[],
   address: string
 ): Promise<ContractData[]> {
