+++ v2/perps-v2/ui/src/index.tsx
@@ -5,17 +5,19 @@ import { ChakraProvider, extendTheme } from '@chakra-ui/react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import Account from './pages/Account';
 import { Header } from './components/Header';
-// import { Actions } from './Actions';
 import { createRoot } from 'react-dom/client';
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client';
 import { PERPS_V2_DASHBOARD_GRAPH_URL } from './utils/constants';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { resolvers, typeDefs } from './queries/resolved';
 
 const queryClient = new QueryClient();
 
 const client = new ApolloClient({
   uri: PERPS_V2_DASHBOARD_GRAPH_URL,
   cache: new InMemoryCache(),
+  resolvers,
+  typeDefs,
 });
 
 const router = createBrowserRouter([
